Issues #1, #2, #3, #4 and #6. Modified cordova push plugin files to modify the behavior of the push notifications  for Android/iOS devices.
Issue #1. Modified the Kotlin files to add a new behavior in the push notifications: when a critical notification type is received by Firebase then a notification will be always displayed in the phone, no matter the current sound and the DND settings.

Issue #2. Modified the Kotlin files to add the ability to change the DND settings. The user will be redirected to the settings of the phone to manually change them and grant access to the app

Issue #3. Modified the native files to create a local notification that will group the following notifications received on the phone. This local notification will have the data of all received notification. When this local notification is pressed, the app will redirect the user to the Inbox page.

Issue #4. Added logic to remove a specific notification from the OS tray when a reminder is opened/read in the Inbox page.

Issue #6. Modified logic to play custom sounds specified in the payload of the received push notification.